Move ServiceYAML into its own file
